tests: Fix casting soup to str
Casting soup to str is not correct here, because it is compared with a
text object.  Therefore this will fail with Python 2.7 with recent
versions of BeautifulSoup.  (I guess the reason why it worked before was
that older version of BeautifulSoup used to return unicode from the
`__str__` function, which is incorrect behaviour.)